+++ src/my_utilities/wrappers/_rspace.py
@@ -55,6 +55,7 @@ class RSpace():
         
         # check if the variable is scalar: https://stackoverflow.com/questions/38088392/how-do-you-check-for-a-scalar-in-r
         if ro.r['is.atomic'](r_obj)[0] and ro.r['length'](r_obj)[0] == 1:
+            # check in Python side if it is a single-element list/array
             if isinstance(value_rpy, (list, tuple, np.ndarray, )) and len(value_rpy) == 1:
                 # assert len(value_rpy) == 1, f'Unexpected length for a scalar variable: {value_rpy}'
                 return value_rpy[0]